Crate libp2p_pnet

The pnet protocol implements Pre-shared Key Based Private Networks in libp2p, as specified in the spec

Libp2p nodes configured with a pre-shared key can only communicate with other nodes with the same key.

Structs

Fingerprint

A PreSharedKey fingerprint computed from a PreSharedKey

PnetConfig

Private network configuration

PnetOutput

The result of a handshake. This implements AsyncRead and AsyncWrite and can therefore be used as base for additional upgrades.

PreSharedKey

A pre-shared key, consisting of 32 bytes of random data.

Enums

KeyParseError

Error when parsing a PreSharedKey

PnetError

Error when writing or reading private swarms